+++ test/test_helper.exs
@@ -12,6 +12,7 @@ Mimic.copy(File)
 Mimic.copy(IO)
 Mimic.copy(Lightning.FailureEmail)
 Mimic.copy(Lightning.Projects.Provisioner)
+Mimic.copy(Lightning.MetadataService)
 Mimic.copy(Mix.Tasks.Lightning.InstallSchemas)
 
 # Other ExUnit configuration can be found in `config/runtime.exs`,